pyproject.toml
[project]
name = "ml-template"
version = "0.1.0"
description = "A machine learning project template"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Daniel Kiesewalter", email = "dkiesewalter97@gmail.com" }]
keywords = ["machine-learning", "deep-learning", "pytorch", "template"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "deepspeed>=0.16.4",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "lightning>=2.5.1",
    "mlflow>=2.21.3",
    "omegaconf>=2.3.0",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.0",
    "onnxscript>=0.5.6",
    "onnxscript>=0.5.6",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "tensorboard>=2.19.0",
    "torch>=2.7.1",
    "torchmetrics>=1.7.3",
    "torchvision>=0.22.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.19.0",
    "myst-parser>=4.0.1",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
    "sphinxcontrib-napoleon>=0.7",
]

[project.urls]
Homepage = "https://github.com/yourusername/ml-template"
Documentation = "https://ml-template.readthedocs.io"
Repository = "https://github.com/yourusername/ml-template"
Issues = "https://github.com/yourusername/ml-template/issues"

[project.scripts]
ml-train = "ml_template.train:main"
ml-export = "ml_template.export:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ml_template"]

# ===================================
# Ruff Configuration
# ===================================
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG"]  # Allow unused arguments in tests

[tool.ruff.lint.isort]
known-first-party = ["ml_template"]

# ===================================
# MyPy Configuration
# ===================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ===================================
# Pytest Configuration
# ===================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

# ===================================
# Coverage Configuration
# ===================================
[tool.coverage.run]
source = ["src/ml_template"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

# ===================================
# Bandit Configuration
# ===================================
[tool.bandit]
exclude_dirs = ["tests", "docs"]
skips = ["B101"]  # Allow assert in non-test code

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "myst-parser>=4.0.1",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
    "sphinxcontrib-napoleon>=0.7",
]
